Use the mapping interface for `DataHandler`
The core of this commit is in `DataHandler`, the rest is just
refactoring to work with the changes.

This commit implements `__len__`, `__getitem__`, and `__iter__` dunder
methods on `DataHandler`. The first two are just renamings of
`get_nevents` and `get_variable_arr` to standard Python names; the
last returns an iterator over variable names in the dataset, just like
a dict. By implementing these three methods and inheriting from
`collections.abc.Mapping`, `DataHandler` works with `len`, item
access (i.e. `dh[var_name]`), `in`, and more, all for free [1].

To comply with the Python data model, `__getitem__` (formerly
`get_variable_arr`) raises `KeyError` instead of `RuntimeError` for an
invalid variable name.

[1] https://docs.python.org/3/library/collections.abc.html

[2] https://docs.python.org/3/reference/datamodel.html#object.__getitem__